requirements.txt
fastapi
uvicorn
python-multipart
sounddevice
numpy
faster-whisper
pyyaml
pynput
pyperclip
requests
packaging